+++ frontend/packages/dev-console/src/components/topology/index.ts
@@ -1,5 +1,4 @@
-export { default as getDevConsoleComponentFactory } from './components/devConsoleComponetFactory';
-export { default as getBindableDevConsoleTopologyDataModel } from './dev-console-data-transformer';
+export { getDevConsoleComponentFactory } from './components/devConsoleComponetFactory';
 export {
   providerProvidesServiceBinding,
   providerCreateServiceBinding,
